:repeat: changing name of the class and specify the menu id
